SCSI-IOCS Support
SCSIのHDSイメージからbootできるようになりました。　scsiexrom.dat を./keropi に入れておいてください。
C11、C++11でbuildします。

+++ SDL2/cdaspi.h
@@ -1,24 +1,24 @@
 #ifndef CDASPI_H_
 #define	CDASPI_H_
 
-int CDASPI_Open(void);
+int32_t CDASPI_Open(void);
 void CDASPI_Close(void);
-int CDASPI_Wait(void);
-int CDASPI_ReadTOC(void* buf);
-int CDASPI_Read(long block, BYTE* buf);
-int CDASPI_ExecCmd(BYTE *cdb, int cdb_size, BYTE *out_buff, int out_size);
-int CDASPI_IsOpen(void);
+int32_t CDASPI_Wait(void);
+int32_t CDASPI_ReadTOC(void* buf);
+int32_t CDASPI_Read(int32_t block, uint8_t* buf);
+int32_t CDASPI_ExecCmd(uint8_t *cdb, int32_t cdb_size, uint8_t *out_buff, int32_t out_size);
+int32_t CDASPI_IsOpen(void);
 
 void CDASPI_EnumCD(void);
 
 void CDASPI_Init(void);
 void CDASPI_Cleanup(void);
 
-extern BYTE CDASPI_CDName[8][4];
-extern int CDASPI_CDNum;
+extern uint8_t CDASPI_CDName[8][4];
+extern int32_t CDASPI_CDNum;
 
 extern HINSTANCE ASPIDLL;
-DWORD (*pSendASPI32Command)(LPSRB);
-DWORD (*pGetASPI32SupportInfo)(VOID);
+uint32_t (*pSendASPI32Command)(LPSRB);
+uint32_t (*pGetASPI32SupportInfo)(VOID);
 
 #endif // CDASPI_H_
