+++ SDL2/dswin.h
@@ -3,15 +3,15 @@
 
 #include "common.h"
 
-int DSound_Init(unsigned long rate, unsigned long length);
-int DSound_Cleanup(void);
+int32_t DSound_Init(uint32_t rate, uint32_t length);
+int32_t DSound_Cleanup(void);
 
 void DSound_Play(void);
 void DSound_Stop(void);
-void FASTCALL DSound_Send0(long clock);
+void FASTCALL DSound_Send0(int32_t clock);
 
-void DS_SetVolumeOPM(long vol);
-void DS_SetVolumeADPCM(long vol);
-void DS_SetVolumeMercury(long vol);
+void DS_SetVolumeOPM(int32_t vol);
+void DS_SetVolumeADPCM(int32_t vol);
+void DS_SetVolumeMercury(int32_t vol);
 
 #endif /* dswin_h__ */
