+++ SDL2/winui.h
@@ -3,14 +3,14 @@
 
 #include "common.h"
 
-extern	BYTE	Debug_Text, Debug_Grp, Debug_Sp;
-extern	DWORD	LastClock[4];
+extern	int32_t	Debug_Text, Debug_Grp, Debug_Sp;
+extern	int32_t	LastClock[4];
 
 extern char cur_dir_str[];
-extern int cur_dir_slen;
+extern int32_t cur_dir_slen;
 
 void WinUI_Init(void);
-int WinUI_Menu(int first);
+int32_t WinUI_Menu(int32_t first);
 float WinUI_get_vkscale(void);
 void send_key(void);
 
@@ -25,15 +25,15 @@ struct menu_flist {
 	char name[MFL_MAX][MAX_PATH];
 	char type[MFL_MAX];
 	char dir[4][MAX_PATH];
-	int ptr;
-	int num;
-	int y;
+	int32_t ptr;
+	int32_t num;
+	int32_t y;
 };
 
 extern char menu_item_key[][15];
 extern char menu_items[][15][30];
 
-int WinUI_get_drv_num(int key);
+int32_t WinUI_get_drv_num(int32_t key);
 
 #ifndef _winx68k_gtkui_h
 #define _winx68k_gtkui_h
