+++ x68k/mfp.h
@@ -3,7 +3,7 @@
 
 #include "common.h"
 
-extern	BYTE MFP[24];
+extern	uint8_t MFP[24];
 
 #define MFP_GPIP	0
 #define MFP_AER		1
@@ -30,16 +30,16 @@ extern	BYTE MFP[24];
 #define MFP_TSR		22
 #define MFP_UDR		23
 
-//extern long Timer_Prescaler[8];
-extern long Timer_Count[4];
-extern BYTE LastKey;
-//extern BYTE KeyReadFlag;
+//extern int32_t Timer_Prescaler[8];
+extern int32_t Timer_Count[4];
+extern uint8_t LastKey;
+//extern uint8_t KeyReadFlag;
 
 void MFP_Init(void);
-BYTE FASTCALL MFP_Read(DWORD adr);
-void FASTCALL MFP_Write(DWORD adr, BYTE data);
-void FASTCALL MFP_Timer(long clock);
+uint8_t FASTCALL MFP_Read(int32_t adr);
+void FASTCALL MFP_Write(int32_t adr, uint8_t data);
+void FASTCALL MFP_Timer(int32_t clock);
 void FASTCALL MFP_TimerA(void);
-void MFP_Int(int irq);
+void MFP_Int(int32_t irq);
 
 #endif //_winx68k_mfp
